refactor: crop variety edit
