+++ apps/picsa-apps/dashboard/src/app/modules/crop-information/pages/variety-details/variety-details.component.html
@@ -1,1 +1,38 @@
-<p>variety-details works!</p>
+<div class="page-content">
+  <h2>Edit Details</h2>
+  <form class="form-content" [formGroup]="entryForm" (ngSubmit)="submitForm()">
+    <mat-form-field appearance="outline" floatLabel="always">
+      <mat-label>Crop</mat-label>
+      <input matInput formControlName="crop" style="display: none" />
+      <picsa-form-crop-select formControlName="crop" style="margin-top: 1rem"></picsa-form-crop-select>
+    </mat-form-field>
+    <mat-form-field appearance="outline">
+      <mat-label>Variety</mat-label>
+      <input matInput formControlName="variety" />
+    </mat-form-field>
+    <mat-form-field appearance="outline">
+      <mat-label>Label</mat-label>
+      <input matInput placeholder="E.g. local name" formControlName="label" />
+    </mat-form-field>
+
+    <!-- <div class="form-data">
+      <label for="water_upper">Water Upper:</label>
+      <input id="water_upper" type="number" formControlName="water_upper" />
+    </div>
+    <div class="form-data">
+      <label for="water_lower">Water Lower:</label>
+      <input id="water_lower" type="number" formControlName="water_lower" />
+    </div>
+    <div class="form-data">
+      <label for="length_upper">Length Upper:</label>
+      <input id="length_upper" type="number" formControlName="length_upper" />
+    </div>
+    <div class="form-data">
+      <label for="length_lower">Length Lower:</label>
+      <input id="lenght_lower" type="number" formControlName="length_lower" />
+    </div> -->
+    <button mat-raised-button color="primary" class="submitButton" type="submit" [disabled]="!entryForm.valid">
+      Submit
+    </button>
+  </form>
+</div>
