+++ apps/picsa-apps/dashboard/src/app/modules/crop-information/pages/variety-details/variety-details.component.scss
@@ -1,0 +1,4 @@
+.form-content {
+  display: flex;
+  flex-direction: column;
+}
