+++ apps/picsa-apps/dashboard/src/app/modules/crop-information/pages/variety/variety.component.html
@@ -1,7 +1,7 @@
 <div class="page-content">
   <div style="display: flex; align-items: center">
     <h2 style="flex: 1">Crop Variety</h2>
-    <button mat-stroked-button color="primary" routerLink="entry"><mat-icon>add</mat-icon>Add New Entry</button>
+    <button mat-stroked-button color="primary" routerLink="new"><mat-icon>add</mat-icon>Add New Entry</button>
   </div>
   @if(service.cropProbabilities){
   <picsa-data-table [data]="service.cropProbabilities" [options]="tableOptions"></picsa-data-table>
